Rename helper file, changed old comment
